pyproject.toml
[project]
name = "ai-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "fastapi[standard]>=0.128.7",
    "psycopg[binary]>=3.3.2",
    "pydantic-ai-slim[openai]>=1.57.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
    "PyJWT>=2.11.0",
    "cryptography>=46.0.5",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.0",
]
